// screens/sharedComponents/Interface.ts
export interface MargInterface {
  marginTop: number;
  marginRight: number;
  marginBottom: number;
  marginLeft: number;
}

export interface PaddinInterface {
  paddingTop: number;
  paddingRight: number;
  paddingBottom: number;
  paddingLeft: number;
}

export interface VertInterface {
  marginTop: number;
  marginBottom: number;
  marginVertical: number;
}